fixed initial schema

+++ src/main/scala/playground/Playground.scala
@@ -29,12 +29,12 @@ object Playground extends App {
     */
   val carsSchema = StructType(Array(
     StructField("Name", StringType),
-    StructField("Miles_per_Gallon", IntegerType),
-    StructField("Cylinders", IntegerType),
-    StructField("Displacement", IntegerType),
-    StructField("Horsepower", IntegerType),
-    StructField("Weight_in_lbs", IntegerType),
-    StructField("Acceleration", DoubleType),
+    StructField("Miles_per_Gallon", DoubleType),
+    StructField("Cylinders", LongType),
+    StructField("Displacement", DoubleType),
+    StructField("Horsepower", LongType),
+    StructField("Weight_in_lbs", LongType),
+    StructField("Acceleration", LongType),
     StructField("Year", StringType),
     StructField("Origin", StringType)
   ))
